Changing general structure
* Updating project folder structure
* Adding new extensions and plugins to eslint
* Updating ignored files
