+++ .eslintrc.json
@@ -4,7 +4,14 @@
     "jest": true,
     "es6": true
   },
-  "extends": ["airbnb", "plugin:jest-dom/recommended", "plugin:prettier/recommended"],
+  "extends": [
+    "airbnb",
+    "eslint:recommended",
+    "plugin:react/recommended",
+    "plugin:react-hooks/recommended",
+    "plugin:jest-dom/recommended",
+    "plugin:prettier/recommended"
+  ],
   "globals": {
     "Atomics": "readonly",
     "SharedArrayBuffer": "readonly"
@@ -17,7 +24,7 @@
     "ecmaVersion": 2019,
     "sourceType": "module"
   },
-  "plugins": ["jest-dom", "react"],
+  "plugins": ["jest-dom", "react", "react-hooks"],
   "rules": {
     "arrow-body-style": "off",
     "comma-dangle": ["warn", "only-multiline"],
