+++ .prettierignore
@@ -1,3 +1,4 @@
+build
 dist
 coverage
 node_modules
