+++ src/App/App.css
@@ -1,0 +1,49 @@
+.App {
+  text-align: center;
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+
+.App-link {
+  color: #61dafb;
+}
+
+.smaller {
+  transform: scale(0.9);
+}
+
+.centered {
+  display: flex;
+  justify-content: center;
+}
+
+.form {
+  background: linear-gradient(
+    to right,
+    rgba(239, 239, 239, 100%) 0%,
+    rgba(219, 219, 219, 100%) 50%,
+    rgba(239, 239, 239, 100%) 100%
+  );
+}
+
+.particles {
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  z-index: -1;
+}
+
+.noselect {
+  user-select: none;
+}
