+++ src/components/FaceRecognition/face-recognition.component.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import './face-recognition.styles.css';
+
+const FaceRecognition = ({ imageUrl, box }) => {
+  return (
+    <div className="centered ma">
+      <div className="absolute mt2">
+        <img id="inputimage" alt="" src={imageUrl} width="500px" heigh="auto" />
+        <div
+          className="bounding-box"
+          style={{ top: box.topRow, right: box.rightCol, bottom: box.bottomRow, left: box.leftCol }}
+        />
+      </div>
+    </div>
+  );
+};
+
+export default FaceRecognition;
