+++ src/components/FaceRecognition/index.js
@@ -1,0 +1,3 @@
+import FaceRecognition from './face-recognition.component';
+
+export default FaceRecognition;
