+++ src/components/ImageLinkForm/image-link-form.styles.css
@@ -1,0 +1,4 @@
+.custom-shadow-w {
+  width: 700px;
+  box-shadow: 0px 0px 4px 4px rgba(0, 0, 0, 0.3);
+}
