+++ src/components/ImageLinkForm/index.js
@@ -1,0 +1,3 @@
+import ImageLinkForm from './image-link-form.component';
+
+export default ImageLinkForm;
