+++ src/components/Logo/index.js
@@ -1,0 +1,3 @@
+import Logo from './logo.component';
+
+export default Logo;
