+++ src/components/Navigation/index.js
@@ -1,0 +1,3 @@
+import Navigation from './navigation.component';
+
+export default Navigation;
