+++ src/components/Navigation/navigation.styles.css
@@ -1,0 +1,5 @@
+.navigation-bar {
+  display: flex;
+  justify-content: flex-end;
+  font-weight: bold;
+}
