+++ src/components/Rank/index.js
@@ -1,0 +1,3 @@
+import Rank from './rank.component';
+
+export default Rank;
