+++ src/components/Rank/rank.component.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Rank = ({ name, entries }) => {
+  return (
+    <div>
+      <div className="white f4">{`Hello ${name}, your current entry count is...`}</div>
+      <div className="white f2">{entries}</div>
+    </div>
+  );
+};
+
+export default Rank;
