+++ src/components/Register/index.js
@@ -1,0 +1,3 @@
+import Register from './register.component';
+
+export default Register;
