--- src/components/css/FaceRecognition.css
@@ -1,16 +1,0 @@
-.bounding-box {
-    position: absolute;
-    box-shadow: inset 0 0 0 3px #149df2;
-    display: -ms-flexbox;
-    display: flex;
-    -ms-flex-wrap: wrap;
-    flex-wrap: wrap;
-    -ms-flex-pack: center;
-    justify-content: center;
-    cursor: pointer;	
-}
-
-.frame {
-	/*border: 2vh solid white;*/
-	margin-bottom:  30px;
-}
