--- src/components/css/ImageLinkForm.css
@@ -1,4 +1,0 @@
-.custom-shadow-w {
-	width: 700px;
-	box-shadow:0px 0px 4px 4px rgba(0, 0, 0, 0.3);
-}
