--- src/components/css/Navigation.css
@@ -1,5 +1,0 @@
-.navigation-bar {
-	display: flex;
-	justify-content: flex-end;
-	font-weight: bold;
-}
