--- src/components/js/Rank.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-const Rank = ({name, entries}) => {
-  return (
-    <div>
-      <div className='white f4'>
-        {`Hello ${name}, your current entry count is...`}
-      </div>
-      <div className='white f2'>
-        {entries}
-      </div>
-    </div>
-  );
-}
-
-export default Rank;
