+++ src/index.css
@@ -1,18 +1,17 @@
 body {
-	margin: 0;
-	padding: 0;
-	font-family: "Courier-New", "Courier", monospace;
-	-webkit-font-smoothing: antialiased;
-	-moz-osx-font-smoothing: grayscale;
-  	background-image: radial-gradient(circle, rgb(255, 140, 20) 20%, rgb(150, 90, 250) 90%);
-    background-attachment: fixed;
+  margin: 0;
+  padding: 0;
+  font-family: 'Courier-New', 'Courier', monospace;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  background-image: radial-gradient(circle, rgb(255, 140, 20) 20%, rgb(150, 90, 250) 90%);
+  background-attachment: fixed;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
 }
 
 button {
-	cursor: pointer;
+  cursor: pointer;
 }
